update stable branch to match latest stable release
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/branches/stable@30473 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Foundation/NSProxy.h
@@ -35,9 +35,9 @@ extern "C" {
 @interface NSProxy <NSObject>
 {
 @public
-    Class		isa;
+    Class	isa;
 @private
-    unsigned int	_retain_count;
+    NSUInteger	_retain_count;
 }
 
 + (id) alloc;
@@ -55,7 +55,7 @@ extern "C" {
 + (void) release;
 + (BOOL) respondsToSelector: (SEL)aSelector;
 + (id) retain;
-+ (unsigned int) retainCount;
++ (NSUInteger) retainCount;
 
 - (id) autorelease;
 - (Class) class;
@@ -63,7 +63,7 @@ extern "C" {
 - (void) dealloc;
 - (NSString*) description;
 - (void) forwardInvocation: (NSInvocation*)anInvocation;
-- (unsigned int) hash;
+- (NSUInteger) hash;
 - (id) init;
 - (BOOL) isEqual: (id)anObject;
 - (BOOL) isKindOfClass: (Class)aClass;
@@ -73,7 +73,7 @@ extern "C" {
 - (void) release;
 - (BOOL) respondsToSelector: (SEL)aSelector;
 - (id) retain;
-- (unsigned int) retainCount;
+- (NSUInteger) retainCount;
 - (id) self;
 - (Class) superclass;
 - (NSZone*) zone;
